kkkkkkkk

+++ Jack.Application/FamiliaApp.cs
@@ -152,7 +152,7 @@ namespace Jack.Application
                 familia = Find(ID);
                 dtoFamilia = new DTOFamilia(familia.Codigo, familia.Nome, familia.IsSacolinha,
                                             familia.IsConsistente, familia.Contato, familia.Nivel,
-                                            familia.Status, familia.DataAtualizacao);
+                                            (int)familia.Status, familia.DataAtualizacao);
             }
             catch (Exception ex)
             {
